api listarTareas, crear tarea, eliminar tarea, detalle pregunta

+++ DESARROLLO/SAV/BACKEND/sav-backend/src/main/java/pe/edu/savbackend/dao/UsuarioDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import pe.edu.savbackend.domain.usuario.UsuarioDto;
 import pe.edu.savbackend.entity.Usuario;
 
-public interface UsuarioDao extends CrudRepository<Usuario, Long>{
+public interface UsuarioDao extends CrudRepository<Usuario, Integer>{
 	
 	public Usuario findByUsername(String username);
 	
@@ -16,6 +16,9 @@ public interface UsuarioDao extends CrudRepository<Usuario, Long>{
 			+ " join Persona p on p.idUsuario = u.id "
 			+ " where u.username = ?1")
 	UsuarioDto buscarPorUsername(String username);
+	
+//	@Query(value = "select max(u.id) + 1 from Usuario u")
+//    Integer nextId();
 }
